feat: mainTab과 view 파일 분리 (#13)
